animosic_backend/requirements.txt
fastapi==0.115.12
uvicorn==0.31.0
pandas==2.2.3
spotipy==2.24.0
joblib==1.4.2
python-dotenv==1.0.1
scikit-learn==1.6.1